Split tests into some files
